Update bignum.h
It appears the compilation issue is related to the use of BN_check_prime in the isPrime method of CBigNum (found in src/bignum.h). The function BN_check_prime is not available in all versions of OpenSSL. On modern Ubuntu distributions with newer OpenSSL versions, the equivalent function is usually BN_is_prime_ex.

To fix the issue, you should replace BN_check_prime with BN_is_prime_ex. The updated code in bignum.h would look like this: